pyproject.toml
[project]
name = "rbac-version-2"
version = "0.1.0"
description = "A comprehensive RBAC (Role-Based Access Control) system built with FastAPI and PostgreSQL"
readme = "README.md"
authors = [
    { name = "Gulab Ahmad", email = "gulabahmad724@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "sqlalchemy>=2.0.42",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "uvicorn[standard]>=0.27.0",
    "python-dotenv>=1.1.1",
    "email-validator>=2.2.0",
    "requests>=2.32.4",
]

[project.scripts]
rbac-version-2 = "rbac_version_2:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
